% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Extract_TC_GEP}
\alias{Extract_TC_GEP}
\title{Extracting Tumor Cell-based GEP}
\usage{
Extract_TC_GEP(x)
}
\arguments{
\item{x, y}{GEP of tumor and normal tissues}
}
\value{
The output from \code{\link{print}}
}
\description{
Extracting Tumor Cell-based GEP
}
\examples{
Extract_TC_GEP(x,y)
\dontrun{
Extract_TC_GEP(x,y)
}

}
