% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{MetaPhenotype}
\alias{MetaPhenotype}
\title{Title}
\usage{
MetaPhenotype()
}
\arguments{
\item{data}{GEP of tumor cells}
}
\value{
tumor cell-based metabolic phenotype of each sample
}
\description{
Title
}
