% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documents.R
\docType{data}
\name{normal}
\alias{normal}
\title{Non-cancerous breast tissue microarray data}
\format{
A data frame:
\describe{
\item{row}{gene symbol}
\item{column}{sample names}
}
}
\usage{
normal
}
\description{
5 non-cancerous breast tissue microarray data, downloaded from GSE87455
}
\keyword{datasets}
