% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documents.R
\docType{data}
\name{tumor}
\alias{tumor}
\title{BRCA microarray data}
\format{
A data frame:
\describe{
\item{row}{gene symbol}
\item{column}{sample names}
}
}
\usage{
tumor
}
\description{
30 BRCA bulk microarray data, downloaded from GSE87455
}
\keyword{datasets}
