% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_TC_GEP.R
\name{Extract_TC_GEP}
\alias{Extract_TC_GEP}
\title{Extracting Tumor Cell-based GEP}
\usage{
Extract_TC_GEP(tumor, normal)
}
\arguments{
\item{tumor}{Bulk GEP of tumor tissues}

\item{normal}{Bulk GEP of adjacent normal tissues, or post-treatment tissues with pathologic stage pT0N0}
}
\value{
GEP of tumor cells
}
\description{
Extracting Tumor Cell-based GEP
}
