% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaPhenotype.R
\name{MetaPhenotype}
\alias{MetaPhenotype}
\title{Classify tumor into three metabolic phenotype for seven pathways}
\usage{
MetaPhenotype(TC, pvalue)
}
\arguments{
\item{pvalue}{}
}
\description{
Classify tumor into three metabolic phenotype for seven pathways
}
